Tyoe of car ID changed to long

+++ src/main/java/com/koro/carshomework3/exception/CarAlreadyExistsException.java
@@ -2,7 +2,7 @@ package com.koro.carshomework3.exception;
 
 public class CarAlreadyExistsException extends RuntimeException {
 
-    public CarAlreadyExistsException(int id) {
+    public CarAlreadyExistsException(long id) {
         super("Car with id=" + id + " already exists");
     }
 }
