+++ src/main/java/com/koro/carshomework3/exception/CarNotFoundException.java
@@ -2,7 +2,7 @@ package com.koro.carshomework3.exception;
 
 public class CarNotFoundException extends RuntimeException {
 
-    public CarNotFoundException(int id) {
+    public CarNotFoundException(long id) {
         super("Could not find car with id: " + id);
     }
 }
